+++ test/Pantry/TypesSpec.hs
@@ -10,6 +10,7 @@ module Pantry.TypesSpec
     ( spec
     ) where
 
+import           Data.Aeson.WarningParser ( WithJSONWarnings (..) )
 import qualified Data.ByteString.Char8 as S8
 import qualified Data.Yaml as Yaml
 import           Distribution.Types.PackageName ( mkPackageName )
@@ -22,7 +23,6 @@ import           Pantry.Internal
                    ( Tree (..), TreeEntry (..), mkSafeFilePath, parseTree
                    , renderTree
                    )
-import           Pantry.Internal.AesonExtended
 import qualified Pantry.SHA256 as SHA256
 import           RIO
 import qualified RIO.Text as T